it works

+++ lib/screens/home/components/popular_product.dart
@@ -23,8 +23,8 @@ class PopularProducts extends StatelessWidget {
                 colors: [],
                 categorieId: element["categorieId"],
                 title: element["titre"],
-                oldPrice: element.data()['oldPrix'].toDouble(),
-                price: element["prix"].toDouble(),
+                oldPrice: element.data()['oldPrix'],
+                price: element["prix"],
                 description: element["description"]));
 
             //print("****************");
