+++ lib/screens/produitCategories/categories_produits_liste_screen.dart
@@ -46,8 +46,8 @@ class CategoriesLsiteProduits extends StatelessWidget {
                                   categorieId: element["categorieId"],
                                   title: element["titre"],
                                   countlikes: element["countlikes"],
-                                  oldPrice: element.data()['oldPrix'].toDouble(),
-                                  price: element["prix"].toDouble(),
+                                  oldPrice: element.data()['oldPrix'],
+                                  price: element["prix"],
                                   description: element["description"]));
                   }
